factors out library dependencies and upgrades to library v2 format [ch4779]
