+++ CHANGELOG.md
@@ -1,0 +1,6 @@
+# Changes for the InternetButton Library
+
+## 0.1.11
+
+- library migration to v2 format
+- NeoPixel and ADXL362 dependencies have been factored out to external libraries
