+++ examples/2_Blink_ALL_the_LEDs/2_Blink_ALL_the_LEDs.cpp
@@ -1,4 +1,4 @@
-#include "InternetButton/InternetButton.h"
+#include "InternetButton.h"
 
 /* This InternetButton library has some useful functions.
 Here we blink ALL the LEDs instead of just one.*/
