+++ examples/3_Buttons_and_LEDs/3_Buttons_and_LEDs.cpp
@@ -1,4 +1,4 @@
-#include "InternetButton/InternetButton.h"
+#include "InternetButton.h"
 
 /* How about we make this interactive? */
