+++ examples/5_Motion/5_Motion.cpp
@@ -1,4 +1,4 @@
-#include "InternetButton/InternetButton.h"
+#include "InternetButton.h"
 #include "math.h"
 
 /*Did you know that the InternetButton can detect if it's moving? It's true!
