+++ examples/8_MakingMusic/8_MakingMusic.cpp
@@ -1,6 +1,6 @@
 // You know the routine by now, just going to comment on the new things!
 
-#include "InternetButton/InternetButton.h"
+#include "InternetButton.h"
 
 InternetButton b = InternetButton();
