+++ library.properties
@@ -1,0 +1,7 @@
+name=InternetButton
+version=0.1.11
+license=LGPL
+author=Richard Whitney <richard@particle.io>
+sentence=Functions to make the Internet Button easy to use! If you have an original SparkButton, make sure to use `begin(1)`
+dependencies.neopixel=0.0.10
+dependencies.particle_ADXL362=0.0.1
