--- spark.json
@@ -1,7 +1,0 @@
-{
-    "name": "InternetButton",
-    "version": "0.1.10",
-    "author": "Richard Whitney <richard@particle.io>",
-    "license": "LGPL",
-    "description": "Functions to make the Internet Button easy to use! If you have an original SparkButton, make sure to use `begin(1)`"
-}
